Refactor API structure and add cheque verification routes
Modularized the API by separating database configuration, middleware, routes, and service logic. Added cheque verification endpoint under /api/v1/cheque, implemented request logging middleware, and improved database connection management for better maintainability and scalability.
